requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Document processing
PyMuPDF==1.23.8
python-docx==1.1.0
unstructured[pdf]==0.11.2
pytesseract==0.3.10
pdf2image==1.16.3
camelot-py[cv]==0.11.0
pandas==2.1.3
openpyxl==3.1.2

# AI and NLP
openai>=1.40.0  # Required for structured outputs with beta.chat.completions.parse
langchain==0.0.348
tiktoken==0.5.2
faiss-cpu==1.7.4
chromadb==0.4.18

# Database
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Utils
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
